+++ free/src/main/scala/cats/free/FreeApplicative.scala
