+++ free/src/main/scala/cats/free/Inject.scala
