+++ free/src/main/scala/cats/free/package.scala
