+++ free/src/test/scala/cats/free/CoyonedaTests.scala
@@ -1,8 +1,8 @@
 package cats
-package tests
+package free
 
+import cats.tests.CatsSuite
 import cats.arrow.NaturalTransformation
-import cats.free.Coyoneda
 import cats.laws.discipline.{FunctorTests, SerializableTests}
 
 import org.scalacheck.Arbitrary
