+++ free/src/test/scala/cats/free/FreeApplicativeTests.scala
@@ -1,8 +1,8 @@
 package cats
-package tests
+package free
 
+import cats.tests.CatsSuite
 import cats.arrow.NaturalTransformation
-import cats.free.FreeApplicative
 import cats.laws.discipline.{CartesianTests, ApplicativeTests, SerializableTests}
 import cats.laws.discipline.eq.{tuple3Eq, tuple2Eq}
 import cats.data.State
