+++ free/src/test/scala/cats/free/FreeTests.scala
@@ -1,8 +1,8 @@
 package cats
-package tests
+package free
 
+import cats.tests.CatsSuite
 import cats.arrow.NaturalTransformation
-import cats.free.{Free, Trampoline}
 import cats.laws.discipline.{CartesianTests, MonadTests, SerializableTests}
 import cats.laws.discipline.eq._
 import cats.laws.discipline.arbitrary.function0Arbitrary
